+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 ### Added
+- Add link to the latest available version ([#31](https://github.com/MechanicalFlower/HazyRoad/pull/31))
 ### Changed
 - Use the SFX bus for foot step sounds ([#30](https://github.com/MechanicalFlower/HazyRoad/pull/30))
 ### Deprecated